Force Bower dependencies to work with grunt-wiredep
Added custom configuration options to `bower.json` allowing
`grunt-wiredep` to automatically inject `<script/>` and CSS
`<link/>` nodes to the HTML file.

This allows the Grunt `wiredep` task to inject the jQuery-Template
`script` tag in HTML files.  Previously, the task was logging
warning messages as it was unable to work out which file should
be included.  As a result we were forced to include the file manually.

Likewise, the styles file required by SwipeBox. It was not marked as
a main file in the package's Bower file so `wiredep` could not inject
it into `index.html`.