Integrate Swipebox via Grunt
When you execute `grunt serve:dist` Swipebox icons are
placed in a custom path and they are correctly referened
from within the corresponding CSS file.

However, the solution is a bit sub-optimal at the moment as the
whole `vendor.css` file is scanned in search of Swipebox image paths
that are being replaced with custom ones.  Another issue is
Swipebox icons are not `rev`-ed like other assets are.
This is because I have not been able to inject these tasks
in the `usemin` context in Grunt.

This is something that needs to be addressed at some point.