Add Grunt symlink and configure it for JS unit tests
To avoid issues with referencing the JavaScript classes source code
they are now symlink-ed under `test/src/`.  This makes it easier
to run unit tests in PhantomJS with the `grunt test` task (this
wouldn't be a problem if you simply wanted to run the index file
located under the test directory).